Incorporating feedback
- Use SceneDelegate
- Update iOS and RN for new schema where name is not optional

If you have run the app before, you will need to delete the app from your simulator to rerun this app.
