+++ swift-ios/Task Tracker/AppDelegate.swift
@@ -7,20 +7,6 @@
 //
 
 import UIKit
-import RealmSwift
-
-struct Constants {
-    // Set this to your Realm App ID found in the Realm UI.
-    static let REALM_APP_ID = "myrealmapp-vjmee"
-    
-    // The base URL points to the server hosting MongoDB Realm.
-    static let REALM_BASE_URL = "https://realm-dev.mongodb.com"
-}
-
-let app = RealmApp(Constants.REALM_APP_ID, configuration: AppConfiguration(baseURL: Constants.REALM_BASE_URL,
-    transport: nil,
-    localAppName: nil,
-    localAppVersion: nil))
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -28,11 +14,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
-
-        window = UIWindow(frame: UIScreen.main.bounds)
-        window?.makeKeyAndVisible()
-        window?.rootViewController = UINavigationController(rootViewController: WelcomeViewController())
-
         return true
     }
