+++ swift-ios/Task Tracker/SceneDelegate.swift
@@ -7,17 +7,35 @@
 //
 
 import UIKit
+import RealmSwift
+
+struct Constants {
+    // Set this to your Realm App ID found in the Realm UI.
+    static let REALM_APP_ID = "myrealmapp-vjmee"
+    
+    // The base URL points to the server hosting MongoDB Realm.
+    static let REALM_BASE_URL = "https://realm-dev.mongodb.com"
+}
+
+let app = RealmApp(Constants.REALM_APP_ID, configuration: AppConfiguration(baseURL: Constants.REALM_BASE_URL,
+    transport: nil,
+    localAppName: nil,
+    localAppVersion: nil))
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
     var window: UIWindow?
 
-
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         // Use this method to optionally configure and attach the UIWindow `window` to the provided UIWindowScene `scene`.
         // If using a storyboard, the `window` property will automatically be initialized and attached to the scene.
         // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
-        guard let _ = (scene as? UIWindowScene) else { return }
+ 
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+               
+        window = UIWindow(windowScene: windowScene)
+        window?.makeKeyAndVisible()
+        window?.rootViewController = UINavigationController(rootViewController: WelcomeViewController())
     }
 
     func sceneDidDisconnect(_ scene: UIScene) {
